requirements.txt
alembic==1.13.1
anthropic==0.25.6
BeautifulSoup4==4.12.3
boto3==1.34.88
bs4==0.0.1
openai==1.23.2
pinecone-client==3.2.2
psycopg2-binary==2.9.9
pyfiglet==1.0.2
pymupdf==1.24.2
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.0
retry==0.9.2
retrying==1.3.4
requests~=2.31.0
sqlalchemy==2.0.30
unidecode==1.3.8

